fixed compiler warnings C4146 C4244 C6297 C26454

+++ src/decoder.c
@@ -158,7 +158,7 @@ static int _get_and_advance_uint32(nanocbor_value_t *cvalue, uint32_t *value,
 {
     uint64_t tmp = 0;
     int res = _get_uint64(cvalue, &tmp, NANOCBOR_SIZE_WORD, type);
-    *value = tmp;
+    *value = (uint32_t)tmp;
 
     return _advance_if(cvalue, res);
 }
@@ -434,7 +434,7 @@ static int _decode_float(nanocbor_value_t *cvalue, float *value)
     int res
         = _get_uint64(cvalue, &tmp, NANOCBOR_SIZE_WORD, NANOCBOR_TYPE_FLOAT);
     if (res == 1 + sizeof(uint32_t)) {
-        uint32_t ifloat = tmp;
+        uint32_t ifloat = (uint32_t)tmp;
         memcpy(value, &ifloat, sizeof(uint32_t));
         return _advance_if(cvalue, res);
     }
